chore(package): add yoshi-stylable-dependencies
required to kick in yoshi stylable v3 configuration.

this dependency already comes from wix-ui-core. however, for users of
wix-ui-tpa that might not be enough, because yoshi-stylable-dependencies
would be considered a transitive dependency. because of that it might
get nested into
`node_modules/wix-ui-core/node_modules/yoshi-stylable-dependencies`.

dependencies defined in yoshi-stylable-dependencies might
also get nested, creating a mix between stylable versions.

to avoid such scenario it is better to set yoshi-stylable-dependencies
as dependency of wix-ui-tpa